chore: Replace html files with handlebars
changed .html ext with .hbs for all html files. the handlebars files will be compiled back into html. the reason behind this is when we create a rev-manifest.json that maps original files to their fingerprinted counterparts using the gulp-rev module we will be able to have the correct file paths in the generated html files.